Added support for Tx cache store writes
Modified CacheWriterInterceptor to wrap writes to the cache store in a local transaction. If the cache store support Tx (JDBC) then it will ensure all of the writes made by the node to the store will be atomic which currently they are not.

This combined with  a non-distributed/shared cache configuration should ensure the store and the cache stay in sync.